github-ai-tracker/pyproject.toml
[project]
name = "github-ai-tracker"
version = "1.0.0"
description = "GitHub AI/Vibe Coding 趨勢追蹤器 - 自動爬取 AI 開發工具相關的熱門開源專案並同步到 Notion"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "notion-client>=2.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
